// src/api/enums/ProposalType.ts
// Copyright (c) 2017-2018, The Particl Market developers
// Distributed under the GPL software license, see the accompanying
// file COPYING or https://github.com/particl/particl-market/blob/develop/LICENSE

/**
 * ProposalType
 *
 */

export enum ProposalType {

    PUBLIC_VOTE = 'PUBLIC_VOTE',
    ITEM_VOTE = 'ITEM_VOTE'

}